chore: more more

+++ scripts/trantbox/arch-mirrorlist-integration.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
-if [ ! -f /etc/pacman.d/mirrorlist.old ]; then
-    mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist.old
-fi
+if [ -f /run/host/etc/pacman.d/mirrorlist ] &&
+    [ -f /etc/pacman.d/mirrorlist ] &&
+    [ ! -f /etc/pacman.d/mirrorlist.old ]; then
 
-if [ ! -f /etc/pacman.d/mirrorlist ]; then
+    mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist.old
     ln -sf /run/host/etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist
 fi
